GameOverできるように。
